[RENAME] repository -> jpa repository
