+++ clickpick/src/main/java/com/clickpick/jparepository/CommentJpaRepository.java
@@ -1,4 +1,4 @@
-package com.clickpick.repository;
+package com.clickpick.jparepository;
 
 import com.clickpick.domain.Comment;
 import jakarta.persistence.EntityManager;
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Repository
 @RequiredArgsConstructor
-public class CommentRepository {
+public class CommentJpaRepository {
     private final EntityManager em;
 
     // 댓글 저장
