+++ clickpick/src/main/java/com/clickpick/jparepository/NoticeJpaRepository.java
@@ -1,4 +1,4 @@
-package com.clickpick.repository;
+package com.clickpick.jparepository;
 
 import com.clickpick.domain.Notice;
 import jakarta.persistence.EntityManager;
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Repository
 @RequiredArgsConstructor
-public class NoticeRepository {
+public class NoticeJpaRepository {
     private final EntityManager em;
 
     // 공지 등록
