+++ clickpick/src/main/java/com/clickpick/service/UserService.java
@@ -1,6 +1,6 @@
 package com.clickpick.service;
 
-import com.clickpick.repository.UserRepository;
+import com.clickpick.jparepository.UserJpaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -9,10 +9,10 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class UserService {
 
-    private final UserRepository userRepository;
+    private final UserJpaRepository userRepository;
 
     @Autowired
-    public UserService(UserRepository userRepository) {
+    public UserService(UserJpaRepository userRepository) {
         this.userRepository = userRepository;
     }
